+++ CHANGELOG.md
@@ -6,6 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.0] - 2020-03-03
+
+### Added
+
+- Field options `optionsModel()` and `optionsLabel()`
+
+### Removed
+
+- Field option `model()`
+
 ## [1.0.0-alpha.2] - 2020-02-27
 
 ### Changed